style(two_hop): fix review comments
